Revert "Generate "process synched data" code" Unreviewed build fix. rdar://140125116
286671@main introduced build failures in certain build configurations.
Reverting while the cause is investigated.

This reverts commit 286671@main (376f4eaf9750f0df5b0f91ab0dad9e81fe217c37).

Canonical link: https://commits.webkit.org/286734@main
