+++ Source/WebCore/Sources.txt
@@ -1,4 +1,4 @@
-// Copyright (C) 2017-2024 Apple Inc. All rights reserved.
+// Copyright (C) 2017-2023 Apple Inc. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions
@@ -3365,7 +3365,6 @@ JSMathMLElementWrapperFactory.cpp
 JSSVGElementWrapperFactory.cpp
 MathMLElementFactory.cpp
 PlugInsResourcesData.cpp
-ProcessSyncClient.cpp
 SVGElementFactory.cpp
 StyleBuilderGenerated.cpp
 StylePropertyShorthandFunctions.cpp
